bentoml/requirements.txt
bentoml
torch
Pillow
numpy
ultralytics